requirements.txt
langchain
pydantic
streamlit
beautifulsoup4
requests
sqlite-utils
langchain_google_genai
pyyaml
apscheduler
pytest
altair
python-dotenv
